requirements.txt
fastapi
uvicorn
python-dotenv
PyJWT
bcrypt
python-multipart
sqlalchemy
pydantic[email]
